+++ gem/lib/pagy/modules/abilities/configurable.rb
@@ -3,12 +3,6 @@
 class Pagy
   # Configuration methods
   module Configurable
-    # Deprecated: Sync the pagy JavaScript targets. Use sync(:javascripts, ...) instead
-    def sync_javascript(...)
-      warn "[PAGY] 'Pagy.sync_javascript(...)' is deprecated: use 'Pagy.sync(:javascript, ...)' instead."
-      sync(:javascript, ...)
-    end
-
     # Sync the pagy resource targets.
     def sync(resource, destination, *targets)
       files    = ROOT.join("#{resource}s").glob("{#{targets.join(',')}}")